+++ reacher_agents/replay_buffers.py
@@ -87,7 +87,7 @@ class ReplayBuffer(Buffer):
             torch.from_numpy(
                 np.vstack([e.action for e in experiences if e is not None])
             )
-            .long()
+            .float()
             .to(self.device)
         )
         rewards = (
